Add files via upload
Set literal's as variables
If a code block was done more then once then I created a function for it
I added scripting to send each of the AI's the challenge question, then asked each of the AI's to rank the different challenge questions and to get a short explanation of why
Asked each of the AI's to answer the first challenge question by OPENAI. Then ran a compare of the response by the different AI's on all AI's for a ranking and short explanation why.
My plan was to do this final step with all AI's but started getting lost in the code. 
I think a possible better set up would be use an Array of Hashs for the different AI's question/response/challenge. <br>I do not think Cursor (or Notebook) would be the best option for that.
I did not test with GROQ or with OLLAMA